Optimisation de l'affichage

+++ taxi-dermistes-web/src/js/clients.js
@@ -1,5 +1,8 @@
 window.onload = function() { // Attend que la page termine de charger
 
+    var countError = 0
+    var messagesAlert = []
+
     function afficherRes(res) {
         $('#res').html('')
         $('#res').html(res)
@@ -9,23 +12,23 @@ window.onload = function() { // Attend que la page termine de charger
         return JSON.parse(res.responseText)
     }
 
-    function setStatusForm(statut, nomchamp) {
-        var champForm = $("[name='" + nomchamp + "']")
+    function setStatusForm(statut, nomChamp, idForm) {
+        var champForm = $("#" + idForm + " [name='" + nomChamp + "']")
         var divParent = champForm.parent()
 
         if (statut == "default") {
             divParent.removeClass("has-success has-error has-danger")
             divParent.find(".form-control-feedback").remove()
         } else if (statut == "success") {
-            setStatusForm("default", nomchamp)
+            setStatusForm("default", nomChamp, idForm)
             divParent.addClass("has-success has-feedback")
             divParent.append('<span class="glyphicon glyphicon-ok form-control-feedback" aria-hidden="true"></span>')
         } else if (statut == "error") {
-            setStatusForm("default", nomchamp)
+            setStatusForm("default", nomChamp, idForm)
             divParent.addClass("has-error has-feedback")
             divParent.append('<span class="glyphicon glyphicon-remove form-control-feedback" aria-hidden="true"></span>')
         } else if (statut == "warning") {
-            setStatusForm("default", nomchamp)
+            setStatusForm("default", nomChamp, idForm)
             divParent.addClass("has-warning has-feedback")
             divParent.append('<span class="glyphicon glyphicon-warning-sign form-control-feedback" aria-hidden="true"></span>')
         }
@@ -34,33 +37,36 @@ window.onload = function() { // Attend que la page termine de charger
 
     function setStatusFormAll(statut, idForm) {
         var form = formToJSON($("#" + idForm))
-        for (var nomchamp in form) {
-            setStatusForm(statut, nomchamp)
+        for (var nomChamp in form) {
+            setStatusForm(statut, nomChamp, idForm)
         }
     }
 
-    function setMainAlert(statut, messages) {
+    function setAlert(statut, messages, id) {
         var str = "<ul>"
         for (var i = 0; i < messages.length; i++) {
             str += "<li>" + messages[i] + "</li>"
         }
         str += "</ul>"
-        var mainAlert = $("#mainAlert")
-        var mainAlertDiv = $("#mainAlert div")
-        mainAlertDiv.html("");
-        mainAlertDiv.html(str);
+        var alert = $("#" + id)
+        var alertMessage = $("#" + id + " div")
+        alertMessage.html("");
+        alertMessage.html(str);
         if (statut == "default") {
-            mainAlert.removeClass("hidden alert-warning alert-danger alert-success")
-            mainAlert.addClass("hidden")
+            alert.removeClass("hidden alert-warning alert-danger alert-success")
+            alert.addClass("hidden")
         } else if (statut == "success") {
-            mainAlert.removeClass("hidden alert-warning alert-danger alert-success")
-            mainAlert.addClass("alert-success")
+            alert.removeClass("hidden alert-warning alert-danger alert-success")
+            alert.addClass("alert-success")
         } else if (statut == "error") {
-            mainAlert.removeClass("hidden alert-warning alert-danger alert-success")
-            mainAlert.addClass("alert-danger")
+            alert.removeClass("hidden alert-warning alert-danger alert-success")
+            alert.addClass("alert-danger")
         } else if (statut == "warning") {
-            mainAlert.removeClass("hidden alert-warning alert-danger alert-success")
-            mainAlert.addClass("alert-warning")
+            alert.removeClass("hidden alert-warning alert-danger alert-success")
+            alert.addClass("alert-warning")
+
+        } else if (statut == "nocolor") {
+            alert.removeClass("hidden alert-warning alert-danger alert-success")
         }
     }
 
@@ -75,9 +81,8 @@ window.onload = function() { // Attend que la page termine de charger
         return indexed_array
     }
 
-    function afficheTab(res) {
-        $('#res').html('')
-        var tab = "<table class='table table-hover'><tr><th>id</th><th>nom</th><th>email</th></tr>";
+    function creerTabClient(res) {
+        var tab = "<table class='table table-hover table-condensed table-bordered'><tr><th>id</th><th>nom</th><th>email</th></tr>";
         for (client in res) {
             /*for (info in res[client]) {
                 tab = "<td>" + res[client].info + "</td>" + tab
@@ -90,6 +95,12 @@ window.onload = function() { // Attend que la page termine de charger
             tab += "<td>" + res[client].client_mail + "</td></tr>"
         }
         tab += "</table>"
+        return tab
+    }
+
+    function afficheTab(res) {
+        $('#res').html('')
+        var tab = creerTabClient
         $('#res').append(tab)
     }
 
@@ -112,57 +123,94 @@ window.onload = function() { // Attend que la page termine de charger
 
     $('#creer').on('submit', function(e) {
         e.preventDefault()
+        var idForm = "creer"
         var formData = formToJSON($(this))
+
+        //ETAPE 1 : Vérifications côté client
         if (!validator.isEmpty(formData.client_name)) {
-            if (!validator.isEmail(formData.client_mail)) {
-                afficherRes('Adresse email incorrecte')
-                setStatusForm('error', "client_mail")
-                setMainAlert('error', ["Adresse email incorrecte"])
-            } else {
-                setStatusForm('success', "client_mail")
-                $.ajax({
-                    url: 'http://localhost:3001/api/clients',
-                    type: 'POST',
-                    dataType: 'json', // On désire recevoir du JSON
-                    contentType: 'application/json; charset=utf-8',
-                    data: JSON.stringify(formData),
-                    success: function(res, statut) {
-                        if (res.status == 'success') {
-                            afficherRes('Client ajouté. Identifiant : ' + res.id)
-                            setStatusFormAll("success", "creer")
-                            setMainAlert('success', ["Votre inscription a bien été enregistrée !"])
-                        }
-                    },
-                    error: function(res, statut, erreur) {
-                        afficherRes('Erreur : ' + getError(res).message)
-                        setMainAlert('error', [getError(res).message])
-                    }
-                })
-            }
+            setStatusForm('success', "client_name", idForm)
+
         } else {
-            setStatusForm('error', "client_name")
-            setMainAlert('error', ["Vous devez spécifier un nom"])
+            setStatusForm('error', "client_name", idForm)
+            messagesAlert.push("Vous devez spécifier un nom")
+            countError += 1
         }
+        if (!validator.isEmail(formData.client_mail)) {
+            countError += 1
+            afficherRes('Adresse email incorrecte')
+            setStatusForm('error', "client_mail", idForm)
+            messagesAlert.push("Adresse email incorrecte")
+        } else {
+            setStatusForm('success', "client_mail", idForm)
+        }
+
+        //ETAPE 2 : Envoi des données au serveur
+        if (countError == 0) {
+            $.ajax({
+                url: 'http://localhost:3001/api/clients',
+                type: 'POST',
+                dataType: 'json', // On désire recevoir du JSON
+                contentType: 'application/json; charset=utf-8',
+                data: JSON.stringify(formData),
+                success: function(res, statut) {
+                    if (res.status == 'success') {
+                        afficherRes('Client ajouté. Identifiant : ' + res.id)
+                        setStatusFormAll("success", idForm)
+                        setAlert('success', ["Votre inscription a bien été enregistrée ! <i class='glyphicon glyphicon-ok-circle'></i>", "Votre identifiant est " + res.id], "creerClientAlert")
+                    }
+                },
+                error: function(res, statut, erreur) {
+                    afficherRes('Erreur : ' + getError(res).message)
+                    setAlert('error', [getError(res).message], "creerClientAlert")
+                }
+            })
+        }
+
+        //ETAPE 3 : Affichage des erreurs côté client
+        else {
+            setAlert('error', messagesAlert, "creerClientAlert")
+            countError = 0 //Réinitialisation des erreurs
+            messagesAlert = [] //Réinitialisation des messages
+        }
+
     })
 
     $('#chercher').on('submit', function(e) {
         e.preventDefault()
-        var client_id = parseInt($('#chercher :input').val())
-            //$.getJSON("http://localhost:3000/clients", req, function(data) {
-            //    $("#res").text(JSON.stringify(data));
-            //});
-        $.ajax({
-            dataType: 'json',
-            type: 'GET',
-            url: 'http://localhost:3001/api/clients/' + client_id,
-            success: function(res) {
-                console.log(res)
-                afficherRes(formatageJSON(res))
-            },
-            error: function(res, statut, erreur) {
-                afficherRes('Erreur : ' + getError(res).message)
-            }
-        })
+        var idForm = "chercher"
+        var formData = formToJSON($(this))
+
+        if (validator.isNumeric(formData.client_id)) {
+            setStatusForm('success', "client_id", idForm)
+        } else {
+            setStatusForm('error', "client_id", idForm)
+            messagesAlert.push("L'identifiant ne doit contenir que des chiffres")
+            countError += 1
+        }
+
+        //$.getJSON("http://localhost:3000/clients", req, function(data) {
+        //    $("#res").text(JSON.stringify(data));
+        //});
+        if (countError == 0) {
+            $.ajax({
+                dataType: 'json',
+                type: 'GET',
+                url: 'http://localhost:3001/api/clients/' + formData.client_id,
+                success: function(res) {
+                    console.log(res)
+                    afficherRes(formatageJSON(res))
+                    setAlert('success', ["Identifiant : " + res.client_id, "Nom : " + res.client_name, "Email : " + res.client_mail], "chercherClientAlert")
+                },
+                error: function(res, statut, erreur) {
+                    setAlert('error', [getError(res).message], "chercherClientAlert")
+                    afficherRes('Erreur : ' + getError(res).message)
+                }
+            })
+        } else {
+            setAlert('error', messagesAlert, "chercherClientAlert")
+            countError = 0 //Réinitialisation des erreurs
+            messagesAlert = [] //Réinitialisation des messages
+        }
     })
 
     $('#lister').on('submit', function(e) {
@@ -172,21 +220,46 @@ window.onload = function() { // Attend que la page termine de charger
             type: 'GET',
             url: 'http://localhost:3001/api/clients/',
             success: function(res) {
+                setAlert('nocolor', [creerTabClient(res)], "listerClientAlert")
                 afficheTab(res)
             },
             error: function(res, statut, erreur) {
-                ('Erreur : ' + getError(res).message)
+                setAlert('error', [getError(res).message], "listerClientAlert")
+                    ('Erreur : ' + getError(res).message)
             }
         })
     })
 
     $('#modifier').on('submit', function(e) {
         e.preventDefault()
-
+        var idForm = "modifier"
         var formData = formToJSON($(this))
-        if (!validator.isEmpty(formData.client_mail) && !validator.isEmail(formData.client_mail)) {
-            afficherRes('Adresse email incorrecte')
+
+        if (validator.isNumeric(formData.client_id)) {
+            setStatusForm('success', "client_id", idForm)
         } else {
+            setStatusForm('error', "client_id", idForm)
+            messagesAlert.push("L'identifiant ne doit contenir que des chiffres")
+            countError += 1
+        }
+        if (!validator.isEmpty(formData.client_name)) {
+            setStatusForm('success', "client_name", idForm)
+
+        } else {
+            setStatusForm('error', "client_name", idForm)
+            messagesAlert.push("Vous devez spécifier un nom")
+            countError += 1
+        }
+        if (!validator.isEmail(formData.client_mail)) {
+            countError += 1
+            afficherRes('Adresse email incorrecte')
+            setStatusForm('error', "client_mail", idForm)
+            messagesAlert.push("Adresse email incorrecte")
+        } else {
+            setStatusForm('success', "client_mail", idForm)
+        }
+
+        if (countError == 0) {
 
             $.ajax({
                 url: 'http://localhost:3001/api/clients',
@@ -196,56 +269,80 @@ window.onload = function() { // Attend que la page termine de charger
                 data: JSON.stringify(formData),
                 success: function(res, statut) {
                     if (res.status == 'success') {
+                        setAlert('success', ["Client modifié"], "modifierClientAlert")
                         afficherRes('Client modifié')
                     }
                 },
                 error: function(res, statut, erreur) {
+                    setAlert('error', [getError(res).message], "modifierClientAlert")
                     afficherRes('Erreur : ' + getError(res).message)
                 }
             })
+        } else {
+            setAlert('error', messagesAlert, "modifierClientAlert")
+            countError = 0 //Réinitialisation des erreurs
+            messagesAlert = [] //Réinitialisation des messages
         }
     })
 
     $('#supprimer').on('submit', function(e) {
         e.preventDefault()
-
         var formData = formToJSON($(this))
-        $.ajax({
-            url: 'http://localhost:3001/api/clients',
-            type: 'DELETE',
-            dataType: 'json', // On désire recevoir du JSON
-            contentType: 'application/json; charset=utf-8',
-            data: JSON.stringify(formData),
-            success: function(res, statut) {
-                if (res.status == 'success') {
-                    afficherRes('Client supprimé')
+        var idForm = "supprimer"
+
+        if (validator.isNumeric(formData.client_id)) {
+            setStatusForm('success', "client_id", idForm)
+        } else {
+            setStatusForm('error', "client_id", idForm)
+            messagesAlert.push("L'identifiant ne doit contenir que des chiffres")
+            countError += 1
+        }
+
+        if (countError == 0) {
+            $.ajax({
+                url: 'http://localhost:3001/api/clients',
+                type: 'DELETE',
+                dataType: 'json', // On désire recevoir du JSON
+                contentType: 'application/json; charset=utf-8',
+                data: JSON.stringify(formData),
+                success: function(res, statut) {
+                    if (res.status == 'success') {
+                        afficherRes('Client supprimé')
+                    setAlert('success', ["Client supprimé"], "supprimerClientAlert")
+                    }
+                },
+                error: function(res, statut, erreur) {
+                    setAlert('error', [getError(res).message], "supprimerClientAlert")
+                    afficherRes('Erreur : ' + getError(res).message)
                 }
-            },
-            error: function(res, statut, erreur) {
-                afficherRes('Erreur : ' + getError(res).message)
-            }
-        })
+            })
+        } else {
+            setAlert('error', messagesAlert, "supprimerClientAlert")
+            countError = 0 //Réinitialisation des erreurs
+            messagesAlert = [] //Réinitialisation des messages
+        }
     })
 
 
     $('#reserver').on('submit', function(e) {
         e.preventDefault()
-
         var formData = formToJSON($(this))
+
+
         $.ajax({
-          url: 'http://localhost:3001/api/courses',
-          type: 'POST',
-          dataType: 'json', // On désire recevoir du JSON
-          contentType: 'application/json; charset=utf-8',
-          data: JSON.stringify(formData),
-          success: function(res, statut) {
-            if (res.status == 'success') {
-              afficherRes('Course ajoutée. Identifiant : ' + res.id)
+            url: 'http://localhost:3001/api/courses',
+            type: 'POST',
+            dataType: 'json', // On désire recevoir du JSON
+            contentType: 'application/json; charset=utf-8',
+            data: JSON.stringify(formData),
+            success: function(res, statut) {
+                if (res.status == 'success') {
+                    afficherRes('Course ajoutée. Identifiant : ' + res.id)
+                }
+            },
+            error: function(res, statut, erreur) {
+                afficherRes('Erreur : ' + getError(res).message)
             }
-          },
-          error: function(res, statut, erreur) {
-            afficherRes('Erreur : ' + getError(res).message)
-          }
         })
     })
